backend/apps/agents/runner/Copywriter/requirements.txt
django>=4.2
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
